test: split tests into unit/integration, one class per module
